categoryService/requirements.txt
fastapi
uvicorn
mysql-connector-python
# fastapi
# uvicorn
# gunicorn
# mysql-connector-python